+++ CHANGELOG.md
@@ -10,8 +10,12 @@ Change log
   Remove `liquid-haskell` support.
 * [#10](https://github.com/kowainik/universum/issues/10):
   Remove `Universum.VarArg` module.
+* [#9](https://github.com/kowainik/universum/issues/9):
+  Remove `safe-exceptions` from dependencies. Reexport `Exception` and
+  `SomeException` from `Control.Exception` instead.
 
-universum uses [PVP Versioning][1].
+
+`universum` uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
 [1]: https://pvp.haskell.org